Combining the questionnaire manager and user journey manager
