+++ app/questionnaire/create_questionnaire_manager.py
@@ -1,39 +1,47 @@
-from app.questionnaire_state.user_journey_manager import UserJourneyManager
-from app.routing.routing_engine import RoutingEngine
 from app.questionnaire.questionnaire_manager import QuestionnaireManager
 from app.main import errors
 from app.metadata.metadata_store import MetaDataStore
+from app.parser.schema_parser_factory import SchemaParserFactory
+from app.schema_loader.schema_loader import load_schema
 from flask_login import current_user
 import logging
-from app.main.views.root import load_and_parse_schema
 
 logger = logging.getLogger(__name__)
 
 
 def create_questionnaire_manager():
 
-    metadata = MetaDataStore.get_instance(current_user)
+    questionnaire_manager = QuestionnaireManager.get_instance()
+    if not questionnaire_manager:
+        metadata = MetaDataStore.get_instance(current_user)
 
-    eq_id = metadata.eq_id
-    form_type = metadata.form_type
+        eq_id = metadata.eq_id
+        form_type = metadata.form_type
 
-    logger.debug("Requested questionnaire %s for form type %s", eq_id, form_type)
+        logger.debug("Requested questionnaire %s for form type %s", eq_id, form_type)
 
-    schema = load_and_parse_schema(eq_id, form_type)
-    if not schema:
-        return errors.page_not_found()
-
-    user_journey_manager = UserJourneyManager.get_instance()
-    if not user_journey_manager:
+        schema = load_and_parse_schema(eq_id, form_type)
+        if not schema:
+            return errors.page_not_found()
         logger.debug("Constructing brand new User Journey Manager")
-        user_journey_manager = UserJourneyManager.new_instance(schema)
+        questionnaire_manager = QuestionnaireManager.new_instance(schema)
 
-    # Create the routing engine
-    routing_engine = RoutingEngine(schema)
-
-    # instantiate the questionnaire manager
-    questionnaire_manager = QuestionnaireManager(schema,
-                                                 user_journey_manager,
-                                                 routing_engine,
-                                                 metadata)
     return questionnaire_manager
+
+
+def load_and_parse_schema(eq_id, form_type):
+    """
+    Use the schema loader to get the schema from disk. Then use the parse to construct the object schema
+    :param eq_id: the id of the questionnaire
+    :param form_type: the form type
+    :return: an object schema
+    """
+    # load the schema
+
+    json_schema = load_schema(eq_id, form_type)
+    if json_schema:
+        parser = SchemaParserFactory.create_parser(json_schema)
+        schema = parser.parse()
+        return schema
+    else:
+        return None
