+++ app/questionnaire/state_manager.py
@@ -9,6 +9,12 @@ STATE = "state"
 
 
 class StateManager(object):
+    '''
+    This class is responsible for saving the state of the User Journey Manager into the database.
+    It does this by pickling the Python object graph into JSON and storing that. This means that code
+    changes between deployments can cause odd behaviour if the database isn't wipe, as when it deserializes
+    state from old python code to new python code attributes can be missing.
+    '''
     @staticmethod
     def has_state():
         if settings.EQ_SERVER_SIDE_STORAGE_TYPE == 'DATABASE':
