+++ app/questionnaire/user_action_processor.py
@@ -1,5 +1,6 @@
 from flask_login import current_user
 from app.submitter.submitter import SubmitterFactory
+from app.metadata.metadata_store import MetaDataStore
 from app import settings
 import logging
 
@@ -8,17 +9,17 @@ logger = logging.getLogger(__name__)
 
 class UserActionProcessor(object):
 
-    def __init__(self, schema, metadata, user_journey_manager):
+    def __init__(self, schema, questionnaire_manager):
         self._schema = schema
-        self._metadata = metadata
-        self._user_journey_manager = user_journey_manager
+        self._metadata = MetaDataStore.get_instance(current_user)
+        self._questionnaire_manager = questionnaire_manager
         self._user_actions = self._build_user_action_chain()
 
     def _build_user_action_chain(self):
         # builds the chain of responsibility
         start_questionnaire = StartQuestionnaire(self._schema, self._metadata)
         save_continue = SaveContinue(self._schema, self._metadata)
-        submit = SubmitAnswers(self._schema, self._metadata, self._user_journey_manager)
+        submit = SubmitAnswers(self._schema, self._metadata, self._questionnaire_manager)
         start_questionnaire.set_next_action(save_continue).set_next_action(submit)
         return start_questionnaire
 
@@ -76,14 +77,14 @@ class SaveContinue(UserAction):
 
 
 class SubmitAnswers(UserAction):
-    def __init__(self, schema, metadata, user_journey_manager):
+    def __init__(self, schema, metadata, questionnaire_manager):
         super().__init__(schema, metadata)
         self.action = 'submit_answers'
-        self._user_journey_manager = user_journey_manager
+        self._questionnaire_manager = questionnaire_manager
 
     def perform_action(self):
-        answers = self._user_journey_manager.get_answers()
+        answers = self._questionnaire_manager.get_answers()
         submitter = SubmitterFactory.get_submitter()
         submitted_at = submitter.send_answers(current_user, self._metadata, self._schema, answers)
         logger.debug("setting submitted at %s", submitted_at)
-        self._user_journey_manager.submitted_at = submitted_at.strftime(settings.DISPLAY_DATETIME_FORMAT)
+        self._questionnaire_manager.submitted_at = submitted_at.strftime(settings.DISPLAY_DATETIME_FORMAT)
