+++ app/questionnaire_state/item.py
@@ -13,15 +13,7 @@ class Item(object):
         for child in self.children:
             child.update_state(user_input, schema_item)
 
-    def get_answer(self, id):
-        answers = self.get_answers()
-        if id in answers:
-            return answers[id]
-        else:
-            return None
-
     def get_answers(self):
-        # TODO we can optimise this away once we remove the answer store
         answers = []
         for child in self.children:
             answers.extend(child.get_answers())
